requirements.txt
altair==5.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dateparser==1.2.0
dnspython==2.6.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
gitdb==4.0.11
GitPython==3.1.43
GoogleNews==1.6.14
idna==3.7
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
lxml==5.2.2
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
py4j==0.10.9.7
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.7.3
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.35.0
tenacity==8.4.1
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
watchdog==4.0.1
